backend-python/requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
jsonify==0.5
Flask-Cors==4.0.0